DOME: RF001 -  Criar conta de utilizador user.js moved into obj classes folder updated the imports to match file location updated the code to fix bugs closeRF001 -  Criar conta de utilizador #2
